split file VideoGraphicsView from LineDrawingDialog
