+++ EnvConfig.cpp
@@ -1,4 +1,5 @@
 #include "EnvConfig.h"
+
 #include <QFile>
 #include <QTextStream>
 #include <QDebug>
