+++ VideoStreamWidget.cpp
@@ -1,5 +1,6 @@
 #include "VideoStreamWidget.h"
 #include "CustomMessageBox.h"
+
 #include <QVBoxLayout>
 #include <QHBoxLayout>
 #include <QApplication>
